add: call out examples
